+++ src/styles/carousel.module.css
